fix test import issue
Not sure exactly what was happening but guessing circular import that vite didn't like
