+++ src/routes/Onboarding/CmdK.tsx
@@ -1,5 +1,6 @@
 import usePlatform from 'hooks/usePlatform'
-import { OnboardingButtons, onboardingPaths, useDismiss, useNextClick } from '.'
+import { OnboardingButtons, useDismiss, useNextClick } from '.'
+import { onboardingPaths } from 'routes/Onboarding/paths'
 import { useStore } from '../../useStore'
 
 export default function CmdK() {
