+++ src/routes/Onboarding/paths.ts
@@ -1,0 +1,15 @@
+export const onboardingPaths = {
+  INDEX: '/',
+  CAMERA: '/camera',
+  STREAMING: '/streaming',
+  EDITOR: '/editor',
+  PARAMETRIC_MODELING: '/parametric-modeling',
+  INTERACTIVE_NUMBERS: '/interactive-numbers',
+  COMMAND_K: '/command-k',
+  USER_MENU: '/user-menu',
+  PROJECT_MENU: '/project-menu',
+  EXPORT: '/export',
+  MOVE: '/move',
+  SKETCHING: '/sketching',
+  FUTURE_WORK: '/future-work',
+}
